requirements.txt
fastapi
uvicorn

sqlalchemy
asyncpg
psycopg2-binary

elasticsearch
elasticsearch-dsl

alembic
pydantic
pydantic-settings

python-dotenv
